+++ src/app/button-with-options/terra-button-with-options.component.ts
@@ -16,16 +16,43 @@ import { TerraButtonComponent } from '../button/terra-button.component';
 export class TerraButtonWithOptionsComponent
 {
     // terra button inputs
+    /**
+     * @description Set the caption.
+     * */
     @Input() inputCaption:string;
+    /**
+     * @description Set an icon (e.g. icon-save).
+     * */
     @Input() inputIcon:string;
+    /**
+     * @description If true, the button gets the primary color blue. Default false.
+     * */
     @Input() inputIsPrimary:boolean;
+    /**
+     * @description If true, the button gets the secondary color red. Default false.
+     * */
     @Input() inputIsSecondary:boolean;
+    /**
+     * @description If true, the button gets the tertiary color green. Default false.
+     * */
     @Input() inputIsTertiary:boolean;
+    /**
+     * @description If true, the button will be disabled. Default false.
+     * */
     @Input() inputIsDisabled:boolean;
+    /**
+     * @description Set the tooltip.
+     * */
     @Input() inputTooltipText:string;
+    /**
+     * @description Set the tooltip placement (bottom, top, left, right). Default top.
+     * */
     @Input() inputTooltipPlacement:string; //top, bottom, left, right
 
     // new inputs
+    /**
+     * @description Set an array of buttons which will be shown as a menu with options. Use the TerraButtonInterface to set the buttons of the menu.
+     * */
     @Input() inputOptions:Array<TerraButtonInterface>;
 
     // view children