+++ src/app/dynamic-module-loader/terra-dynamic-module-loader.component.ts
@@ -45,7 +45,7 @@ export class TerraDynamicModuleLoaderComponent implements AfterViewInit, OnChang
 
     public ngOnChanges(changes:SimpleChanges):void
     {
-        if ( changes.hasOwnProperty('inputInputs') )
+        if(changes.hasOwnProperty('inputInputs'))
         {
             this.assignInputProperties();
         }
@@ -98,6 +98,7 @@ export class TerraDynamicModuleLoaderComponent implements AfterViewInit, OnChang
                     }
                 }
             );
+            this._cmpRef.changeDetectorRef.detectChanges();
         }
     }
 }