+++ src/app/forms/input/number-input/terra-number-input.component.ts
@@ -25,7 +25,9 @@ export class TerraNumberInputComponent extends TerraInputComponent
     {
         super(TerraRegex.NUMERIC);
     }
-
+    /**
+     * @deprecated inputValue is deprecated. It will be removed in one of the upcoming releases. Please use ngModel instead.
+     * */
     @Input()
     public set inputValue(v:number)
     {