+++ src/app/terra-components.module.ts
@@ -14,6 +14,7 @@ import {
     ModalModule,
     TooltipModule
 } from 'ngx-bootstrap';
+import { QuillModule } from 'ngx-quill';
 import { TerraComponentsComponent } from './terra-components.component';
 import { TerraTextInputComponent } from './forms/input/text-input/terra-text-input.component';
 import { TerraNumberInputComponent } from './forms/input/number-input/terra-number-input.component';
@@ -66,15 +67,18 @@ import { TerraSplitViewComponent } from './split-view/terra-split-view.component
 import { CommonModule } from '@angular/common';
 import { TerraDynamicComponentLoaderComponent } from './dynamic-component-loader/terra-dynamic-component-loader.component';
 import { BrowserModule } from '@angular/platform-browser';
-import { TerraFileBrowserComponent } from "./file-browser/terra-file-browser.component";
 import { TerraFileInputComponent } from "./forms/input/file-input/terra-file-input.component";
 import { TerraFrontendStorageService } from "./file-browser/terra-frontend-storage.service";
 import { BrowserAnimationsModule } from "@angular/platform-browser/animations";
 import { TerraColorPickerComponent } from "./forms/input/color-picker/terra-color-picker.component";
 import { TerraInteractModule } from "./interactables/interact.module";
 import { TerraSliderComponent } from "./forms/slider/terra-slider.component";
-import { TerraNoResultNoticeComponent} from './no-result/terra-no-result-notice.component';
+import { TerraNoResultNoticeComponent } from './no-result/terra-no-result-notice.component';
 import { TerraButtonWithOptionsComponent } from './button-with-options/terra-button-with-options.component';
+import { FixedHeaderDirective } from './table/fixed-header/fixed-header.directive';
+import { TerraFileBrowserService } from './file-browser/terra-file-browser.service';
+import { TerraFileBrowserComponent } from './file-browser/terra-file-browser.component';
+import { TerraNoteEditorComponent } from './note-editor/terra-note-editor.component';
 import { TerraNoteComponent } from './note/terra-note.component';
 
 export { TerraAlertPanelComponent } from './alert/terra-alert-panel.component';
@@ -143,7 +147,7 @@ export { TerraTextAreaInputComponent } from './forms/input/text-area-input/terra
 export { TerraCardComponent } from './card/terra-card.component';
 export { TerraSyntaxEditorComponent } from './editor/syntax/terra-syntax-editor.component';
 export { TerraSyntaxEditorData } from './editor/syntax/data/terra-syntax-editor.data';
-export { TerraNoResultNoticeComponent} from './no-result/terra-no-result-notice.component';
+export { TerraNoResultNoticeComponent } from './no-result/terra-no-result-notice.component';
 
 @NgModule({
     declarations:    [
@@ -194,6 +198,8 @@ export { TerraNoResultNoticeComponent} from './no-result/terra-no-result-notice.
         TerraFileInputComponent,
         TerraNoResultNoticeComponent,
         TerraButtonWithOptionsComponent,
+        FixedHeaderDirective,
+        TerraNoteEditorComponent,
         TerraNoteComponent
     ],
     entryComponents: [
@@ -239,6 +245,7 @@ export { TerraNoResultNoticeComponent} from './no-result/terra-no-result-notice.
         TerraFileBrowserComponent,
         TerraFileInputComponent,
         TerraButtonWithOptionsComponent,
+        TerraNoteEditorComponent,
         TerraNoteComponent
     ],
     exports:         [
@@ -288,6 +295,8 @@ export { TerraNoResultNoticeComponent} from './no-result/terra-no-result-notice.
         TerraFileInputComponent,
         TerraNoResultNoticeComponent,
         TerraButtonWithOptionsComponent,
+        FixedHeaderDirective,
+        TerraNoteEditorComponent,
         TerraNoteComponent
     ],
     imports:         [
@@ -304,12 +313,14 @@ export { TerraNoResultNoticeComponent} from './no-result/terra-no-result-notice.
         TranslationModule.forRoot(),
         MyDatePickerModule,
         AceEditorModule,
-        TerraInteractModule
+        TerraInteractModule,
+        QuillModule
     ],
     providers:       [
         COMPILER_PROVIDERS,
         TerraNavigatorSplitViewConfig,
         TerraFrontendStorageService,
+        TerraFileBrowserService,
     ],
     bootstrap:       [
         TerraComponentsComponent
@@ -327,6 +338,7 @@ export class TerraComponentsModule
                         TerraBaseService,
                         TerraNavigatorSplitViewConfig,
                         TerraUrlParamsDecorator,
+                        TerraFrontendStorageService,
                         TerraAlertComponent]
         };
     }
@@ -341,6 +353,7 @@ export class TerraComponentsModule
                         TerraBaseService,
                         TerraNavigatorSplitViewConfig,
                         TerraUrlParamsDecorator,
+                        TerraFrontendStorageService,
                         TerraAlertComponent]
         };
     }